2 fixes for 10.5sdk on mac
git-svn-id: http://svn.openframeworks.cc/openFrameworks/trunk/openFrameworks@241 5aee0c02-c622-0410-bcf2-d8652e674aa7

+++ app/ofAppRunner.cpp
@@ -234,7 +234,7 @@ void ofSetVerticalSync(bool bSync) {
 	//--------------------------------------
 #ifdef TARGET_OSX
 	//--------------------------------------
-	long sync = bSync == true ? 1 : 0;
+		GLint sync = bSync == true ? 1 : 0;
 	CGLSetParameter (CGLGetCurrentContext(), kCGLCPSwapInterval, &sync);
 	//--------------------------------------
 #endif
